refactor(orchestrator): extract plan-injection from pre-loop driver
Audit PR (KJC-PCS-0033 follow-up): runPreLoopStages was 496 LOC
("HIGH long-function" finding, biggest in the codebase). Extract the
189-LOC plan-injection block into a dedicated phase module so the
driver shrinks toward orchestration-only.

Result: runPreLoopStages 496 → 318 LOC. Still big, but no longer
the worst offender; further phase extractions can follow the same
pattern in subsequent PRs (intent / brain decision / domain curator
/ auto-HU are the next candidates).

What moved to pre-loop-phases/inject-loaded-plan.js:
  - loadPlan + plan:loaded event + stageResults stamping
  - tests-first gate (refuse run with HUs missing acceptance_tests)
  - PR-E auto-certify pending HUs
  - PR-E assertPlanRunnable anomaly guard
  - planToHuBatch + --hu filter (PR4)
  - PR1 setLiveStatusUpdater + PR3 setLiveOutcomeUpdater wiring
  - syncResultsToPlan callback + plan.status="running"
  - plannedTask construction from plan steps/approach

Behaviour identical — verified by 4037/4037 parent suite, including
the existing plan-to-run integration test (updated to look at the
new file path: pre-loop.js delegates, inject-loaded-plan.js owns
the logic).

The new module structures the long block into small named helpers
(assertEveryHuHasTests, autoCertifyAndAssertRunnable,
applyHuFilterIfRequested, wirePlanCallbacks,
buildPlannedTaskFromLoadedPlan) that can be unit-tested without
spinning up the whole orchestrator. Header documents the
mutator-routing contract preserved from the inlined version.
